Split CLI utils.cpp into more parts
Add base58 encoding/decoding CLI

Use decrypt_or_random in pk_decrypt
